:white_check_mark: Ajout test d'integration + ajout de la logique dans le lib
